[GraphQL] Introducing PackageCache
This data source handles random access queries for packages, initially
to calculate type layouts, but eventually to serve all requests that
are related to package information (modules, structs, functions,
types, layouts, signatures, etc).  It is **upgrade aware**, meaning it
takes into account a package's linkage and type origin tables to
relocate the package IDs of structs that it references.

The store is implemented as an LRU cache on top of an underlying data
source, which currently calls out to the existing RPC using
`SuiClient`, but will eventually read from the DB -- the underlying
data source is abstracted behind a trait (mainly for testability), but
this should also make it simple to adopt the DB when the time comes.

The cache already supports "Reinterpreting Package IDs", an extension
feature that allows a type to be referred to at any package ID after
the one that it was first defined at, because it was easier to add
support for it than prevent the behaviour.

Although this implementation is currently in the `sui-graphql-rpc`
trait, it is self-contained (it only depends on core/underlying
crates) and could also be re-used to do upgrade-aware type and layout
calculation in the indexer codebase.

This PR also integrates an implementation the PackageCache into our
GraphQL server, to serve type layout calculates (necessary to return
the layout of a type and the contents of a move object).

The following related work will be handled in follow-up PRs, to keep
the size of this PR as small as possible.

- Moving the cache's configuration into `ServiceConfig`.
- Implementations of `OpenMoveType`, and its `OpenMoveTypeSignature`
  in GraphQL.

There are also some further TODOs highlighting potential future
optimisations that are not planned for the immediate future, but that
we can revisit if we need to.

Test Plan:

New unit tests:

```
sui-graphql-rpc$ cargo nextest run -- package_cache
```